+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - 0.8.4
   * Add optional argument for delay in milliseconds to button actions `okay` and `bummer`. 
+  * Resetting a django-formset removes all just added sibling collections and unmarks all
+    collections for removal.
 
 - 0.8.3
   * Fix: For ``field_css_classes``, fall back to form name rather than its prefix.